@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/chat-bot",
-  "version": "1.37.0-alpha.0",
+  "version": "1.40.0-alpha.0",
   "description": "",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
@@ -27,10 +27,10 @@
     "yarn": "1"
   },
   "dependencies": {
-    "@hoprnet/hopr-core": "1.37.0-alpha.0",
-    "@hoprnet/hopr-core-connector-interface": "1.37.0-alpha.0",
-    "@hoprnet/hopr-core-ethereum": "1.37.0-alpha.0",
-    "@hoprnet/hopr-utils": "1.37.0-alpha.0",
+    "@hoprnet/hopr-core": "1.40.0-alpha.0",
+    "@hoprnet/hopr-core-connector-interface": "1.40.0-alpha.0",
+    "@hoprnet/hopr-core-ethereum": "1.40.0-alpha.0",
+    "@hoprnet/hopr-utils": "1.40.0-alpha.0",
     "@hoprnet/twitter-api-client": "^0.1.8",
     "@linkdrop/sdk": "1.1.6",
     "debug": "4.2.0",
@@ -38,7 +38,7 @@
     "dotenv-parse-variables": "^1.0.1",
     "firebase-admin": "^9.1.1",
     "global": "^4.4.0",
-    "peer-id": "0.13",
+    "peer-id": "0.14.2",
     "rimraf": "^3.0.2",
     "tinyurl": "1.1.7",
     "wait-for-stuff": "^1.4.0",
@@ -52,7 +52,7 @@
     "@types/node": "~12",
     "@typescript-eslint/eslint-plugin": "4.0.0",
     "@typescript-eslint/parser": "3.10.1",
-    "eslint": "7.11.0",
+    "eslint": "7.12.1",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-import": "^2.20.1",
     "jest": "26.6.1",
