@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/hopr-server",
-  "version": "1.37.0-alpha.0",
+  "version": "1.40.0-alpha.0",
   "description": "gRPC Server wrapper around HOPR Node",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
@@ -48,10 +48,10 @@
   },
   "dependencies": {
     "@grpc/proto-loader": "^0.5.4",
-    "@hoprnet/hopr-core": "1.37.0-alpha.0",
-    "@hoprnet/hopr-core-connector-interface": "1.37.0-alpha.0",
-    "@hoprnet/hopr-protos": "1.37.0-alpha.0",
-    "@hoprnet/hopr-utils": "1.37.0-alpha.0",
+    "@hoprnet/hopr-core": "1.40.0-alpha.0",
+    "@hoprnet/hopr-core-connector-interface": "1.40.0-alpha.0",
+    "@hoprnet/hopr-protos": "1.40.0-alpha.0",
+    "@hoprnet/hopr-utils": "1.40.0-alpha.0",
     "@nestjs/common": "^7.0.0",
     "@nestjs/config": "^0.5.0",
     "@nestjs/core": "^7.0.0",
@@ -61,16 +61,15 @@
     "grpc": "^1.24.3",
     "multiaddr": "^7.5.0",
     "parse-ip-port": "^1.0.0",
-    "peer-id": "^0.13.13",
-    "peer-info": "^0.17.5",
+    "peer-id": "0.14.2",
     "read-pkg-up": "^7.0.1",
     "reflect-metadata": "^0.1.13",
     "rlp": "^2.2.6",
     "rxjs": "^6.6.0"
   },
   "devDependencies": {
-    "@hoprnet/hopr-ethereum": "1.37.0-alpha.0",
-    "@hoprnet/hopr-testing": "1.37.0-alpha.0",
+    "@hoprnet/hopr-ethereum": "1.40.0-alpha.0",
+    "@hoprnet/hopr-testing": "1.40.0-alpha.0",
     "@nestjs/cli": "^7.0.0",
     "@nestjs/schematics": "^7.0.0",
     "@nestjs/testing": "^7.0.0",
@@ -82,7 +81,7 @@
     "@typescript-eslint/parser": "3.10.1",
     "copyfiles": "^2.3.0",
     "cross-env": "^7.0.2",
-    "eslint": "7.11.0",
+    "eslint": "7.12.1",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-import": "^2.20.1",
     "jest": "26.6.1",
