@@ -38,9 +38,9 @@ export default class CloseChannel extends AbstractCommand {
       const { status, receipt } = await this.node.closeChannel(peerId)
 
       if (status === 'PENDING') {
-        return `${chalk.green(`Closing channel, receipt: ${styleValue(receipt, 'hash')}`)}}.`
+        return `${chalk.green(`Closing channel, receipt: ${styleValue(receipt, 'hash')}`)}.`
       } else {
-        return `${chalk.green(`Initiated channel closure, receipt: ${styleValue(receipt, 'hash')}`)}}.`
+        return `${chalk.green(`Initiated channel closure, receipt: ${styleValue(receipt, 'hash')}`)}.`
       }
     } catch (err) {
       return styleValue(err.message, 'failure')
@@ -48,10 +48,6 @@ export default class CloseChannel extends AbstractCommand {
   }
 
   async autocomplete(query: string = '', line: string = ''): Promise<AutoCompleteResult> {
-    if (!query) {
-      return [[this.name()], line]
-    }
-
     let peerIdStrings: string[]
     try {
       peerIdStrings = await this.node.paymentChannels.channel.getAll(
