@@ -1,10 +1,42 @@
+import type PeerId from 'peer-id'
 import { AbstractCommand, AutoCompleteResult, GlobalState } from '../abstractCommand'
 import { styleValue, getOptions, checkPeerIdInput } from '../../utils'
 
-export class Routing extends AbstractCommand {
-  // there is actually another option which takes in peerIds
-  private readonly options: GlobalState['routing'][] = ['manual', 'direct']
+const ROUTING_PATH_PREFIX = 'path='
 
+export const options: ReadonlyArray<GlobalState['routing']> = ['manual', 'direct']
+
+/**
+ * Convert a query to a routing path
+ * @param routing
+ * @returns a promise that resolves into an array for peerIds
+ */
+export async function queryToPeerIds(query: string): Promise<PeerId[]> {
+  if (!query.startsWith(ROUTING_PATH_PREFIX)) throw Error("To enter a routing path, use 'path=<peerId>,<peerId>'.")
+  const routingPath = query.slice(ROUTING_PATH_PREFIX.length, query.length)
+  return routingPathToPeerIds(routingPath)
+}
+
+/**
+ * Looks into `state.routing` and parses the specified peerIds
+ * @param routing
+ * @returns a promise that resolves into an array for peerIds
+ */
+export async function routingPathToPeerIds(path: string): Promise<PeerId[]> {
+  return Promise.all(
+    path
+      .split(',')
+      .filter((peerId) => peerId.length > 0)
+      .map(async (peerId) => await checkPeerIdInput(peerId))
+  )
+}
+
+export function getRouting(state: GlobalState): string {
+  if (state.routingPath.length > 0) return state.routingPath.map((peerId) => peerId.toB58String()).join(',')
+  return state.routing
+}
+
+export class Routing extends AbstractCommand {
   public name() {
     return 'routing'
   }
@@ -14,43 +46,48 @@ export class Routing extends AbstractCommand {
   }
 
   public async execute(query: string, state: GlobalState): Promise<string | void> {
-    if (!query) {
-      return styleValue(state.routing, 'highlight')
-    }
+    try {
+      if (!query) return styleValue(getRouting(state), 'highlight')
 
-    // provided peerIds
-    if (query.match(',')) {
-      try {
-        await Promise.all(query.split(',').map(async (peerId) => await checkPeerIdInput(peerId)))
-        state.routing = query
-      } catch (err) {
-        return styleValue(`Invalid option ${query}.`, 'failure')
+      // reset routingPath
+      state.routingPath = []
+
+      if (options.find((o) => o === query)) {
+        state.routing = query as any
+      } else {
+        const peerIds = await queryToPeerIds(query)
+
+        if (peerIds.length === 0) {
+          state.routing = 'direct'
+        } else {
+          state.routing = 'manual'
+          state.routingPath = peerIds
+        }
       }
-    }
-    // provided a mode
-    else {
-      const option = this.options.find((o) => query === o)
-      if (!option) {
-        return styleValue(`Invalid option ${query}.`, 'failure')
+
+      let message = `You have set your “${styleValue(this.name(), 'highlight')}” settings to “${styleValue(
+        state.routing,
+        'highlight'
+      )}”`
+
+      if (state.routingPath.length > 0) {
+        message = `${message} with a path of “${styleValue(getRouting(state), 'highlight')}”`
       }
 
-      state.routing = option
+      return `${message}.`
+    } catch (err) {
+      return styleValue(err.message, 'failure')
     }
-
-    return `You have set your “${styleValue(this.name(), 'highlight')}” settings to “${styleValue(
-      state.routing,
-      'highlight'
-    )}”.`
   }
 
   public async autocomplete(query: string, line: string): Promise<AutoCompleteResult> {
     // nothing provided, just show all options
     if (!query) {
-      return [getOptions(this.options.map((o) => ({ value: styleValue(o, 'highlight') }))), line]
+      return [getOptions(options.map((o) => ({ value: styleValue(o, 'highlight') }))), line]
     }
 
     // matches a option partly, show matches options
-    const matchesPartly = this.options.filter((option) => {
+    const matchesPartly = options.filter((option) => {
       return option.startsWith(query)
     })
 
