@@ -1,4 +1,3 @@
-import type { Handler } from '../../@types/transport'
 import debug from 'debug'
 import pipe from 'it-pipe'
 import chalk from 'chalk'
@@ -8,7 +7,7 @@ import PeerId from 'peer-id'
 import Multiaddr from 'multiaddr'
 import { CrawlResponse, CrawlStatus } from '../../messages'
 import { LibP2P } from '../../'
-import type { Connection } from '../../@types/transport'
+import type { Connection, Handler } from 'libp2p'
 
 const log = debug('hopr-core:crawler')
 const verbose = debug('hopr-core:verbose:crawl-interaction')
@@ -44,9 +43,9 @@ class Crawler implements AbstractInteraction {
         struct = await this.node
           .dialProtocol(counterparty, this.protocols[0], { signal: options.signal })
           .catch(async (_: Error) => {
-            const peerInfo = await this.node.peerRouting.findPeer(counterparty)
+            const { id } = await this.node.peerRouting.findPeer(counterparty)
 
-            return await this.node.dialProtocol(peerInfo, this.protocols[0], { signal: options.signal })
+            return await this.node.dialProtocol(id, this.protocols[0], { signal: options.signal })
           })
       } catch (err) {
         log(`Could not ask node ${counterparty.toB58String()} for other nodes. Error was: ${chalk.red(err.message)}.`)
