@@ -1,6 +1,4 @@
-import PeerInfo from 'peer-info'
 import PeerId from 'peer-id'
-import libp2p from 'libp2p'
 // @ts-ignore
 import TCP = require('libp2p-tcp')
 // @ts-ignore
@@ -9,67 +7,65 @@ import MPLEX = require('libp2p-mplex')
 import SECIO = require('libp2p-secio')
 import { Heartbeat } from './heartbeat'
 import assert from 'assert'
-import Multiaddr from 'multiaddr'
 import { EventEmitter } from 'events'
 import * as constants from '../../constants'
+import { generateLibP2PMock } from '../../test-utils'
 
 // @ts-ignore
 constants.HEARTBEAT_TIMEOUT = 300
 
 describe('check heartbeat mechanism', function () {
   async function generateNode(options?: { timeoutIntentionally?: boolean }) {
-    const node = await libp2p.create({
-      peerInfo: await PeerInfo.create(await PeerId.create({ keyType: 'secp256k1' })),
-      modules: {
-        transport: [TCP],
-        streamMuxer: [MPLEX],
-        connEncryption: [SECIO]
-      }
-    })
-    node.peerInfo.multiaddrs.add(Multiaddr('/ip4/0.0.0.0/tcp/0'))
+    const { node, address } = await generateLibP2PMock()
+
     node.peerRouting.findPeer = (_: PeerId) => Promise.reject(Error('not implemented'))
 
     await node.start()
 
-    node.interactions = {
-      network: {
-        heartbeat: new Heartbeat(node, (remotePeer) => node.network.heartbeat.emit('beat', remotePeer), options)
-      }
-    }
-
-    node.network = {
+    const network = {
       heartbeat: new EventEmitter()
     }
 
-    return node
+    const interactions = {
+      network: {
+        heartbeat: new Heartbeat(node, (remotePeer) => network.heartbeat.emit('beat', remotePeer), options)
+      }
+    }
+
+    return {
+      node,
+      network,
+      interactions,
+      address
+    }
   }
 
   it('should dispatch a heartbeat', async function () {
     const [Alice, Bob] = await Promise.all([generateNode(), generateNode()])
 
-    await Alice.dial(Bob.peerInfo)
+    await Alice.node.dial(Bob.address)
 
     await Promise.all([
       new Promise((resolve) => {
         Bob.network.heartbeat.once('beat', (peerId: PeerId) => {
-          assert(peerId.isEqual(Alice.peerInfo.id), 'connection must come from Alice')
+          assert(peerId.isEqual(Alice.node.peerId), 'connection must come from Alice')
           resolve()
         })
       }),
-      Alice.interactions.network.heartbeat.interact(Bob.peerInfo.id)
+      Alice.interactions.network.heartbeat.interact(Bob.node.peerId)
     ])
 
-    await Promise.all([Alice.stop(), Bob.stop()])
+    await Promise.all([Alice.node.stop(), Bob.node.stop()])
   })
 
   it('should trigger a heartbeat timeout', async function () {
     const [Alice, Bob] = await Promise.all([generateNode(), generateNode({ timeoutIntentionally: true })])
 
-    await Alice.dial(Bob.peerInfo)
+    await Alice.node.dial(Bob.address)
     let errorThrown = false
     let before = Date.now()
     try {
-      await Alice.interactions.network.heartbeat.interact(Bob.peerInfo.id)
+      await Alice.interactions.network.heartbeat.interact(Bob.node.peerId)
     } catch (err) {
       errorThrown = true
     }
@@ -80,6 +76,6 @@ describe('check heartbeat mechanism', function () {
       `Should reach a timeout, ${Date.now() - before} ${constants.HEARTBEAT_TIMEOUT}`
     )
 
-    await Promise.all([Alice.stop(), Bob.stop()])
+    await Promise.all([Alice.node.stop(), Bob.node.stop()])
   })
 })
