@@ -1,7 +1,7 @@
 import { Crawler } from './crawler'
 import { Heartbeat } from './heartbeat'
 import { LibP2P } from '../../'
-import type { Connection } from '../../@types/transport'
+import type { Connection } from 'libp2p'
 import PeerId from 'peer-id'
 
 class NetworkInteractions {
