@@ -4,10 +4,9 @@ import type { Types } from '@hoprnet/hopr-core-connector-interface'
 
 import { PROTOCOL_ONCHAIN_KEY } from '../../constants'
 import type { AbstractInteraction } from '../abstractInteraction'
-import type PeerInfo from 'peer-info'
 import type PeerId from 'peer-id'
 
-import type { Handler } from '../../@types/transport'
+import type { Handler } from 'libp2p'
 
 import chalk from 'chalk'
 
@@ -31,7 +30,7 @@ class OnChainKey<Chain extends HoprCoreConnector> implements AbstractInteraction
       struct = await this.node._libp2p.dialProtocol(counterparty, this.protocols[0]).catch(async (_: Error) => {
         return this.node._libp2p.peerRouting
           .findPeer(counterparty)
-          .then((peerInfo: PeerInfo) => this.node._libp2p.dialProtocol(peerInfo, this.protocols[0]))
+          .then((peerRoute) => this.node._libp2p.dialProtocol(peerRoute.id, this.protocols[0]))
       })
     } catch (err) {
       throw Error(
