@@ -6,10 +6,9 @@ import pipe from 'it-pipe'
 
 import type { AbstractInteraction } from '../abstractInteraction'
 
-import type { Handler } from '../../@types/transport'
+import type { Handler } from 'libp2p'
 
 import { PROTOCOL_PAYMENT_CHANNEL } from '../../constants'
-import type PeerInfo from 'peer-info'
 import type PeerId from 'peer-id'
 
 class Opening<Chain extends HoprCoreConnector> implements AbstractInteraction {
@@ -34,7 +33,7 @@ class Opening<Chain extends HoprCoreConnector> implements AbstractInteraction {
       struct = await this.node._libp2p.dialProtocol(counterparty, this.protocols[0]).catch(async (_: Error) => {
         return this.node._libp2p.peerRouting
           .findPeer(counterparty)
-          .then((peerInfo: PeerInfo) => this.node._libp2p.dialProtocol(peerInfo, this.protocols[0]))
+          .then((peerRoute) => this.node._libp2p.dialProtocol(peerRoute.id, this.protocols[0]))
       })
     } catch (err) {
       throw Error(
