@@ -1,7 +1,8 @@
 import assert from 'assert'
 import Listener from './listener'
 import Multiaddr from 'multiaddr'
-import { Connection, Upgrader } from '../../@types/transport'
+import { Upgrader } from 'libp2p'
+import type { Connection } from 'libp2p'
 import dgram, { Socket, RemoteInfo } from 'dgram'
 import { handleStunRequest } from './stun'
 import PeerId from 'peer-id'
