@@ -7,7 +7,7 @@ const error = debug('libp2p:tcp:socket:error')
 import toIterable = require('stream-to-it')
 // @ts-ignore
 import toMultiaddr = require('libp2p-utils/src/ip-port-to-multiaddr')
-import { MultiaddrConnection } from '../../@types/transport'
+import { MultiaddrConnection } from 'libp2p'
 import type Multiaddr from 'multiaddr'
 import type { Socket } from 'net'
 
