@@ -2,7 +2,6 @@
 import Hopr from '@hoprnet/hopr-core'
 import type { HoprOptions } from '@hoprnet/hopr-core'
 import type HoprCoreConnector from '@hoprnet/hopr-core-connector-interface'
-import PeerInfo from 'peer-info'
 import PeerId from 'peer-id'
 import Multiaddr from 'multiaddr'
 import debug from 'debug'
@@ -55,7 +54,7 @@ const argv = yargs
   })
   .option('provider', {
     describe: 'A provider url for the Network you specified',
-    default: 'wss://xdai.poanetwork.dev/wss'
+    default: 'wss://ws-mainnet.matic.network'
   })
   .option('host', {
     describe: 'The network host to run the HOPR node on.',
@@ -119,7 +118,7 @@ function parseHosts(): HoprOptions['hosts'] {
 
 async function generateNodeOptions(): Promise<HoprOptions> {
   let options: HoprOptions = {
-    debug: Boolean(process.env.DEBUG),
+    debug: Boolean(process.env.HOPR_DEBUG),
     bootstrapNode: argv.bootstrap,
     network: argv.network,
     bootstrapServers: argv.bootstrap ? [] : [...(await getBootstrapAddresses()).values()],
