@@ -1,9 +1,6 @@
-import PeerInfo from 'peer-info'
 import PeerId from 'peer-id'
 import Multiaddr from 'multiaddr'
 import dns from 'dns'
-
-type ServerMap = Map<string, PeerInfo>
 
 const BOOTSTRAP_ADDRESS = process.env.HOPR_BOOTSTRAP_ADDRESS || '_dnsaddr.bootstrap.basodino.develop.hoprnet.org'
 
@@ -12,10 +9,9 @@ const BOOTSTRAP_ADDRESS = process.env.HOPR_BOOTSTRAP_ADDRESS || '_dnsaddr.bootst
  *   - If there are ENV Variables, use them second
  *   - Otherwise look at the DNS entry for our testnet.
  */
-export async function getBootstrapAddresses(addrs?: string): Promise<ServerMap> {
+export async function getBootstrapAddresses(addrs?: string): Promise<Multiaddr[]> {
   let addresses: string[]
-  let bootstrapServerMap = new Map<string, PeerInfo>()
-  let addr: Multiaddr
+  let servers: Multiaddr[] = []
 
   if (addrs) {
     addresses = addrs.split(',')
@@ -31,17 +27,5 @@ export async function getBootstrapAddresses(addrs?: string): Promise<ServerMap> 
     throw new Error('Bootstrap Addresses could not be found')
   }
 
-  for (let i = 0; i < addresses.length; i++) {
-    addr = Multiaddr(addresses[i].trim())
-
-    let peerInfo = bootstrapServerMap.get(addr.getPeerId())
-    if (peerInfo == null) {
-      peerInfo = await PeerInfo.create(PeerId.createFromB58String(addr.getPeerId()))
-    }
-
-    peerInfo.multiaddrs.add(addr)
-    bootstrapServerMap.set(addr.getPeerId(), peerInfo)
-  }
-
-  return bootstrapServerMap
+  return addresses.map((a: string) => Multiaddr(a.trim()))
 }
