@@ -1,7 +1,9 @@
 import assert from 'assert'
 import PeerId from 'peer-id'
-import { convertPubKeyFromPeerId, convertPubKeyFromB58String } from '.'
-import * as multihashes from 'typestub-multihashes'
+import { convertPubKeyFromPeerId, convertPubKeyFromB58String, hasB58String, getB58String } from '.'
+import { validTweetsMap } from './tweets'
+// @ts-ignore
+import * as multihashes from 'multihashes'
 
 describe(`test convertPubKeyFromPeerId`, function () {
   it(`should equal to a newly created pubkey from PeerId`, async function () {
@@ -14,5 +16,57 @@ describe(`test convertPubKeyFromPeerId`, function () {
     const pubKey = await convertPubKeyFromB58String(testIdB58String)
     const id = PeerId.createFromCID(testIdB58String)
     assert(id.pubKey.toString() === pubKey.toString())
+  })
+})
+
+describe(`test hasB58String`, function () {
+  it(`should return a boolean value`, function () {
+    const response = hasB58String('test')
+    assert(typeof response === 'boolean')
+  })
+  it(`should return false to a content w/o a b58string`, function () {
+    const response = hasB58String('A random string w/o a b58string')
+    assert(response === false)
+  })
+  it(`should return true to a content w/a b58string`, function () {
+    const response = hasB58String('16Uiu2HAmCPgzWWQWNAn2E3UXx1G3CMzxbPfLr1SFzKqnFjDcbdwg')
+    assert(response === true)
+  })
+  it(`should return true to a content w/a b58string`, function () {
+    const tweet = `16Uiu2HAkz2s8kLcY7KTSkQBDUmfD8eSgKVnYRt8dLM36jDgZ5Z7d 
+@hoprnet
+ #HOPRNetwork
+    `
+    const response = hasB58String(tweet)
+    assert(response === true)
+  })
+  it(`should return true for all valid tweets`, function () {
+    assert(validTweetsMap.every(hasB58String) === true)
+  })
+})
+
+describe(`test hasB58String`, function () {
+  it(`should return a string value`, function () {
+    const response = getB58String('test')
+    assert(typeof response === 'string')
+  })
+  it(`should return an empty string to a content w/o a b58string`, function () {
+    const response = getB58String('A random string w/o a b58string')
+    assert(response === '')
+  })
+  it(`should return the b58string to a content w/a b58string`, function () {
+    const response = getB58String('16Uiu2HAmCPgzWWQWNAn2E3UXx1G3CMzxbPfLr1SFzKqnFjDcbdwg')
+    assert(response === '16Uiu2HAmCPgzWWQWNAn2E3UXx1G3CMzxbPfLr1SFzKqnFjDcbdwg')
+  })
+  it(`should return the b58string to a content w/a b58string`, function () {
+    const tweet = `16Uiu2HAkz2s8kLcY7KTSkQBDUmfD8eSgKVnYRt8dLM36jDgZ5Z7d 
+@hoprnet
+ #HOPRNetwork
+    `
+    const response = getB58String(tweet)
+    assert(response === '16Uiu2HAkz2s8kLcY7KTSkQBDUmfD8eSgKVnYRt8dLM36jDgZ5Z7d')
+  })
+  it(`should return a string of 53 characters for all valid tweets`, function () {
+    assert(validTweetsMap.every((tweet) => getB58String(tweet).length === 53) === true)
   })
 })
